test: Moved tests and adapted code
